+++ src/core/tokenizer/mod.rs
@@ -5,7 +5,7 @@ pub(crate) mod cache;
 mod decode;
 mod encode;
 mod error;
-mod patterns;
+pub(crate) mod patterns;
 pub(crate) mod scanner;
 #[cfg(test)]
 mod tests;